Refactored leave application screen into 2 widgets
